Remove a bug causing app to crash whenever comment is added
The reason for this crash is beacuse comment component was trying to
access a prop which was invalid at time of creation of comment field but
is updated once the comment field is created.

To prevent this I rendered the comment conditionally only when commentId
is valid. This solved the issue.